Complete MCP server refactoring to 500-line compliance
- Refactor mcp-server.js from 701 lines to 250 lines (64% reduction)
- Split monolithic file into 3 focused modules:
  * Main server: 250 lines (setup, routing, error handling)
  * ToolHandlers: 340 lines (4 MCP tool implementations)
  * MessageUtils: 140 lines (utilities for tokens, filtering, pagination)
- Maintain identical functionality with improved code organization
- Backup original as mcp-server-original.js for rollback capability

Code Quality Achievement:
✅ 500-line limit compliance
✅ Modular architecture
✅ Single responsibility principle
✅ Full audit trail via git history
